+++ irs-api/src/main/java/org/eclipse/tractusx/irs/services/IrsItemGraphQueryService.java
@@ -139,6 +139,7 @@ public class IrsItemGraphQueryService implements IIrsItemGraphQueryService {
                            .bomLifecycle(bomLifecycle.getLifecycleContextCharacteristicValue())
                            .aspectTypes(aspectTypeValues)
                            .collectAspects(collectAspects)
+                           .callbackUrl(request.getCallbackUrl())
                            .build();
     }